+++ infra/Chain.cpp
@@ -24,7 +24,7 @@ namespace AnalysisTree {
 //    }
 //  }
 //
-//  for (uint i = 1; i < filelists_.size(); i++) {
+//  for (size_t i = 1; i < filelists_.size(); i++) {
 //    this->AddFriend(MakeChain(filelists_.at(i), treenames_.at(i)));
 //  }
 //
@@ -63,7 +63,7 @@ void Chain::InitChain() {
     aliases.emplace_back(LookupAlias(aliases, tree_name));
   }
 
-  for (uint i = 1; i < filelists_.size(); i++) {
+  for (size_t i = 1; i < filelists_.size(); i++) {
     if (aliases.at(i) != treenames_.at(i)) {
       std::cout << "Tree '" << treenames_.at(i) << "' will be friended under the alias '" << aliases.at(i) << "'" << std::endl;
     }
@@ -132,9 +132,9 @@ void Chain::InitConfiguration() {
   std::string name = "Configuration";
   configuration_ = GetObjectFromFileList<Configuration>(filelists_.at(0), name);
 
-  for (uint i = 1; i < filelists_.size(); ++i) {
+  for (size_t i = 1; i < filelists_.size(); ++i) {
     auto* config_i = GetObjectFromFileList<Configuration>(filelists_.at(i), name);
-    for (uint j = 0; j < config_i->GetNumberOfBranches(); ++j) {
+    for (size_t j = 0; j < config_i->GetNumberOfBranches(); ++j) {
       configuration_->AddBranchConfig(config_i->GetBranchConfig(j));
     }
   }
